refactor(editor): receiver-pattern decoration providers with sweeteditor types
EditorDecorationProvider now delivers through an EditorDecorationReceiver
(Accept/IsCancelled), supporting both synchronous and asynchronous
engines. Decoration result and context use SweetEditor core types
directly (StyleSpan, Diagnostic, CodeLensItem, DocumentHighlight,
TextChange) via aliases instead of re-wrapped structs; the conversion
layer is deleted. The demo SweetLine provider is updated accordingly.

+++ examples/preview/src/app.cpp
@@ -414,8 +414,8 @@ huxerui::codeeditor::EditorOptions MakeEditorOptions(
         document.syntax,
         document.text,
         document.key,
-        [breakpoints](uint32_t start_line, uint32_t end_line) {
-          std::vector<std::pair<uint32_t, int32_t>> icons;
+        [breakpoints](size_t start_line, size_t end_line) {
+          std::vector<std::pair<size_t, int32_t>> icons;
           for (uint32_t line : breakpoints.Get()) {
             if (line >= start_line && line <= end_line) {
               icons.emplace_back(line, 2);
